Compressed indirection buffers for avgpool
The indirection buffers for avgpool microkernel is repeated output_height times. Each input pointer in row i+1 points to a constant offset away from the input pointers in row i. We compress the indirection buffer by calculating this offset, and storing it in the compute context. Then the compute function does not need to advance the indirection buffer pointer.

However, we need to handle padding separately, so the indirection buffer is separateed into 3 sections: top, middle, and bottom. The top and bottom are not compressed, and accounts for top and bottom padding. The middle is the compressed section, and if present, will only have a single row of pointers.

Add logic into the compute function to determine which section we are in, which then affects how we calculate the input pointers.

This is applicable for all average pooling microkernel: pixelwise, unipass, multipass.

A new indirection init function is needed because we have to handle the last section differently, we cannot use output_y to calculate the y offset in the indirection buffer to write to.

We keep the current indirection init function around because we still use it in benchmark and in convolution operator.

PiperOrigin-RevId: 566474958