Fix QD8 F32 QC8W IGEMM when SAME padding is used.
The input zero point is not constant and is calculated at runtime. The input is quantized per batch meaning that each batch has a different zero point so multiple zero point buffers are required.

The solution is to fill the zero point buffers before the IGEMM ukernel is called and pass the batch_index as the offset into the zero buffer.

PiperOrigin-RevId: 574836587