In the `f16c` `vsqrt` microkernel, instead of computing the full-precision `sqrt(x)` on the converted `float` values, compute `x * rsqrt(x)` where `rsqrt(x)` is the 12-bit approximation of the reciprocal square root of `x`.
Since the result is converted back to an `f16`, which only has a 10-bit mantissa,
the 12-bit approximation is sufficiently accurate.

Note that we need to be careful when `x` is sufficiently close to zero for
`rsqrt(x)` to return `+Inf`, in which case the result is replaced with `0`. This
allows us to handle both zero and non-finite inputs correctly.

The resulting microkernels are a bit more than 2x faster than the original
versions.

PiperOrigin-RevId: 606600392