Add broadcasting to the Batch Matrix Multiply op.
Broadcasting happens when any of the input batch dimensions is an integer multiple
of the other inputs' dimension, e.g. the batched matrix multiplication two inputs
of the shapes:

 * `A`: `[1, 5, 6, M, K]`, and
 * `B`: `[8, 5, 2, K, N]`,

will result in an output of the shape `[8, 5, 6, M, N]`, i.e.

 * The first batch dimension of `A` is broadcast from `1` to `8`,
 * The second batch dimension of `A` and `B` match at `5`,
 * The third batch dimension of `B` is broadcast from `2` to `6`.

The indices of the inputs to each batch can be computed as the output index modulo
the input dimensions. In the above example, the result

 * `C[4, 1, 3, :, :]`

is the product of the matrices

 * `A[0, 1, 3, :, :]`, and
 * `B[4, 1, 1, :, :]`.

PiperOrigin-RevId: 611110791