Implement `f32-vtanh` microkernels using a 9/6 rational polynomial approximation.
The coefficients of the rational polynomial were computed using a variant of the Remez algorithm to minimize the relative error in the range `[-8, 8]`. The approximation has a maximum relative error of `5.36e-7` or `4.79e-7` (4.5 or 4 ULP), for non-FMA and FMA evaluation, respectively.

Evaluating the rational approximant is ~1.5x faster than the current implementations for `SSE2`, `AVX`, `FMA3`, and `AVX512`.

PiperOrigin-RevId: 627141097