Initialize memory in pack functions that kernels expect to be initialized
This changes packing to not assume that the output has been initialized by a prior memset call. This pattern is a dangerous source of bugs. In addition to that, it requires memsetting the whole pack buffer, which may be less efficient than just initializing memory appropriately in the packing functions.

The case where this is questionable in terms of performance is batch matrix multiply, where packing may be called many times on a buffer that is initialized once. IMO, relying on previous initialization is a risky decision that makes a lot of sketchy assumptions. It would be worth a small performance penalty to avoid that.

Note that most of the small holes don't exist in the pack microkernels, it's mostly just the reference functions that have a lot of small holes.

PiperOrigin-RevId: 735554407