Add `xnn_define_fuse_dims` and `xnn_define_split_dim` subgraph nodes.
`xnn_define_fuse_dims` takes a range of dimensions and replaces them with a single dimension of the size of the product of the dimensions in the range.

`xnn_define_split_dim` takes an axis and replaces it with one or more dimensions, the product of which are equal to the size of the replaced dimension.

PiperOrigin-RevId: 738357192