Pack the LHS of `fully-connected` and `dynamic-fully-connected` ops for one block of rows at a time, in the op itself.
We currently have separate `pack-lh` that produce packed inputs (of type `qd8`, `qp8`, `pqs8`, `pf16`, and `pf32`) that feed into `fully-connected` and `dynamic-fully-connected` ops. This means the entire input tensor is first (converted and) packed, stored in a temporary buffer, and only then read, usually only once, by a single `fully-connected` or `dynamic-fully-connected` op.

This is bad in terms of memory consumption (for the intermediate buffer), cache reuse (the inputs are packed, stored, then re-read later on), and parallel efficiency (synchronization point between the `pack-lh` and `fully-connected` ops).

This change folds the `pack-lh` operations into the `fully-connected` and `dynamic-fully-connected` ops consuming them. The `fully-connected` ops then compute the GEMM in strips of `mr`x`n`, where the `mr`x`k` left hand operand values are first packed in a per-thread buffer and then used immediately.

The per-row LHS packing is gated conditionally on whether the data needed to compute an entire tile row (`mr`x`n`) of the output fit in the L2 cache. For larger inputs, the code reverts to pre-packing the LHS as processing the `n` columns of the output in chunks is usually more cache-efficient.

Internally, the creation and/or inlining of the `pack-lh` ops is now handled as a subgraph optimization, which makes it easier to maintain and extend.

The resulting combined op is **up to 20% faster** than the original two ops, both on a single and multiple threads, when packed `qdint8` or `qduint8` LHSs are used.

On machines with `sme2` support, the packed `pf32` and `pf16` GEMMs are **up to 50% faster** than the original split ops.

Thanks to the gating on the L2 cache size, there are only a few small regressions on matrices with relatively small `k` and/or `n` values (<32).

PiperOrigin-RevId: 753924784